opencl: add support for OpenCL 1.2 subdevices
Summary:
In OpenCL 2.0, clGetContextInfo(CL_CONTEXT_DEVICES) is said to return
"the list of devices and subdevices in context". In OpenCL 1.2, however,
it does not reference subdevices, leading to implementations that do not
include them even if they were provided when creating the context, such
as pocl.

To support those platforms, when an opencl context is attached, if the
device is not found in the context, we search for the root device it
comes from. If this one is found, we manually add the subdevice to
the OpenCV list of devices in the context.

ref T16402

Test Plan: Using KAF subdevices with OpenCV should work

Reviewers: echaloin, mkapfer, etournier, O72 kaf!, O78 kstuff!

Subscribers: mqho

Differential Revision: https://phab.kalray.eu/D13370