opencl: work-around clGetContextInfo not returning subdevices on POCL
Summary:
Even though POCL says in its version call's return it is an OpenCL 2.0,
it does not return subdevices, which OpenCV depends on. A workaround
had been added for this, but it checked specifically for OpenCL 1.2.

This extends it to target POCL too.

ref T19051

Test Plan: CI in kafapps

Reviewers: echaloin, O72 kaf!, O78 kstuff!

Maniphest Tasks: T19051

Differential Revision: https://phab.kalray.eu/D20399